server: update to support authenticated cloning
Enable the ability to use a PAT for authenticated access to GitHub.  In
order to properly support this functionality, switch from gitclone to
simple-git for the git wrapper.  Perform some minor cleanups of sinking
application start up logic into `main`.  Additionally, we perform a
micro-optimization of not fetching the content and checking out the data
until the first service request.

Co-authored-by: Andrew Monshizadeh <amonshiz@thebrowser.company>